move generate_field_summary. more comments.
