+++ test/timeseries_test.py
@@ -1,9 +1,7 @@
 from io import StringIO
 
 from libs.datasets.dataset_utils import AggregationLevel
-from libs.datasets.sources import cds_dataset
 from libs.datasets.timeseries import TimeseriesDataset
-from test.dataset_utils_test import to_dict
 import pandas as pd
 import pytest
 
@@ -13,6 +11,8 @@ pytestmark = pytest.mark.filterwarnings("error")
 
 
 def test_get_subset_and_get_data():
+    # CSV with a unique FIPS value for every region, even countries. In production countries are removed before
+    # TimeseriesDataset is created. A future change may replace FIPS with a more general identifier.
     input_df = pd.read_csv(
         StringIO(
             "city,county,state,fips,country,aggregate_level,date,metric\n"
